When using hypermail -v to print out the config values, MakeConfig() was derefencing the values using a pointer to a long instead of to an int.

+++ src/setup.c
@@ -998,7 +998,7 @@ void MakeConfig(bool comments)
 		if (cfg[i].changed)
 		    /* they differ, show the actual contents */
 		    fprintf(stdout, "%s = %lo\n", cfg[i].label,
-			    *(long *)(cfg[i].value));
+			    *(int *)(cfg[i].value));
 		else if (comments)
 		    /* show the default value, but commented */
 /*
@@ -1011,7 +1011,7 @@ void MakeConfig(bool comments)
 		if (cfg[i].changed)
 		    /* they differ, show the actual contents */
 		    fprintf(stdout, "%s = %ld\n", cfg[i].label,
-			    *(long *)(cfg[i].value));
+			    *(int *)(cfg[i].value));
 		else if (comments)
 		    /* show the default value, but commented */
 /*
